requirements.txt
python-slugify==4.0.0
hdx-python-api==4.1.8
-r docker-requirements.txt